Adding unit tests for the CollectionManager class.
Some light refactoring and renaming as well.
Also redid how the CollectionManager is implemented now, regarding thread safty.

There are now two collections, one for typed collections, one for untyped collections. If no typed collection is present for any modifications, a lock is acquired prior to checking the untyped collection. This same lock is required to convert to a typed collection, so two threads cannot modify the untyped collection at the same time.